# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analytics/events": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Record Analytics Event",
        "description": "Record an aggregate analytics event reported by the browser. No user is stored against the event.",
        "operationId": "record_analytics_event_analytics_events_post",
        "parameters": [
          {
            "name": "x-amzn-oidc-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amzn-Oidc-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsEventRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalyticsEventName": {
        "type": "string",
        "enum": [
          "live_recording_started_or_upload_requested",
          "audio_upload_complete_from_live_recording",
          "audio_upload_complete_from_direct_upload",
          "transcript_requested_for_live_recording",
          "transcript_requested_for_direct_upload",
          "transcript_received_for_live_recording",
          "transcript_received_for_direct_upload",
          "summary_requested",
          "summary_received"
        ],
        "title": "AnalyticsEventName"
      },
      "AnalyticsEventRequest": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/AnalyticsEventName"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AnalyticsEventRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}